+++ changelog.md
@@ -17,7 +17,7 @@
   - Introduce audio volume slider
   - Label for "Screen Controls Position" now consistent with the styles of the rest
 - Internal:
-  - Remove `radio-selection` component
+  - Remove `radio-selection` component and general cleanup
 - Add Valentine's animation
 
 # 0.8.7 Beta - 02/02/2026 13:30 UTC+1