+++ changelog.md
@@ -1,4 +1,4 @@
-# 0.8.8 Beta Valentine's Edition -
+# 0.8.8 Beta Valentine's Edition - 09/02/2026 13:00 UTC+1
 
 - Authentication:
   - Fix bug when date is not defined
@@ -20,6 +20,7 @@
   - Remove `radio-selection` component and general cleanup
   - Update authentication to read `lastGame` from `player` property (API 0.3.2 change)
   - Refactor `FetchRequest`: decouple more from logout logic and add `abort()` method (not used)
+  - Manual image caching for characters
 - Add Valentine's animation
 
 # 0.8.7 Beta - 02/02/2026 13:30 UTC+1